requirements.txt
streamlit
groq
langchain
langchain-groq
langchain-community
python-dotenv
langchain_core



